Switch from pytest-datadir to my own implementation.
It is more flexible regarding the available datadirs,
but does not copy data to a temporary directory. I'm
not yet sure if copying is needed or even a good idea.

Maybe the test functions should do the copying themselves
if they need it instead of relying on the datadir fixture
to do it.