Add support for optional documentation comments
Added a load option to interpret comments after properties as
documentation text for the key defined.  The text is added to
the metadata with the key "_doc".  E.g., the properties file
contents (from the README.rst):

    #: _severity=fatal
    10001=Fatal internal error: %s
    # A fatal internal error occurred.  Please re-run the command
    # with the -D option to generate additional debug information.

defines the "_severity" metadata key and, when "metadoc=True"
is used when loading, the additional key "_doc" with the value

    A fatal internal error occurred.  Please re-run the command
    with the -D option to generate additional debug information.

Signed-off-by: Michael Rohan <mrohan@vmware.com>